requirements.txt
numpy
Pillow
tqdm
scikit-learn
torch
python-mnist
seaborn
matplotlib
pydiffmap
pynndescent
umap-learn